examples/python/requirements.txt
google-cloud-storage==1.32.0
urllib3==1.25.11
